Fix "command not found" at install of ops CLI
The ops CLI failed install via `npm run ops`. This was due to the
missing `bin` property in `package.json`, plus the file system
permissions of the `src/ops.js` file.

In the case of the former, the `bin` property was added to
`package.json` with the path to the `src/ops.js` file. This
exposed the `ops` command to the system.

In the case of the latter, the file was not executable, so the
`ops` command was not found. Setting the file permissions to
`755` fixed the issue.